Extract CLI logic into library and move main tests to tests/unit
