+++ ci/coverage-theoretical-max-receipts/coverage-theoretical-max.json
@@ -3,71 +3,71 @@
   "platforms": {
     "linux": {
       "branches": {
-        "amount": 1796,
+        "amount": 1854,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "functions": {
-        "amount": 2117,
+        "amount": 1932,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "lines": {
-        "amount": 22255,
+        "amount": 20469,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "regions": {
-        "amount": 36948,
+        "amount": 34069,
         "theoretical_max_percent": "100.00000000000000000000000000"
       }
     },
     "macos": {
       "branches": {
-        "amount": 1796,
+        "amount": 1854,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "functions": {
-        "amount": 2117,
+        "amount": 1932,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "lines": {
-        "amount": 22255,
+        "amount": 20469,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "regions": {
-        "amount": 36948,
+        "amount": 34069,
         "theoretical_max_percent": "100.00000000000000000000000000"
       }
     },
     "windows": {
       "branches": {
-        "amount": 1796,
+        "amount": 1854,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "functions": {
-        "amount": 2117,
+        "amount": 1932,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "lines": {
-        "amount": 22255,
+        "amount": 20469,
         "theoretical_max_percent": "100.00000000000000000000000000"
       },
       "regions": {
-        "amount": 36948,
+        "amount": 34069,
         "theoretical_max_percent": "100.00000000000000000000000000"
       }
     }
   },
   "union": {
     "branches": {
-      "amount": 1796
+      "amount": 1854
     },
     "functions": {
-      "amount": 2117
+      "amount": 1932
     },
     "lines": {
-      "amount": 22255
+      "amount": 20469
     },
     "regions": {
-      "amount": 36948
+      "amount": 34069
     }
   }
 }
