+++ ci/llvm-cov-receipts/linux/llvm-cov-report.txt
@@ -46,11 +46,12 @@ bin/mux_client_spawn_probe.rs                 47                 0   100.00%    
 bin/mux_running_version_probe.rs              27                 0   100.00%           1                 0   100.00%          21                 0   100.00%           8                 0   100.00%
 bin/open_pr_flow_probe.rs                    114                 0   100.00%           7                 0   100.00%          64                 0   100.00%          10                 0   100.00%
 bin/tui_smoke.rs                              82                 0   100.00%           2                 0   100.00%          49                 0   100.00%          14                 0   100.00%
+cli.rs                                       918                 0   100.00%          71                 0   100.00%         719                 0   100.00%         124                 0   100.00%
 command.rs                                    55                 0   100.00%           4                 0   100.00%          25                 0   100.00%           4                 0   100.00%
 config/keys.rs                               387                 0   100.00%          21                 0   100.00%         267                 0   100.00%          20                 0   100.00%
 config/mod.rs                                362                 0   100.00%          54                 0   100.00%         250                 0   100.00%          14                 0   100.00%
 lib.rs                                         4                 0   100.00%           1                 0   100.00%           3                 0   100.00%           0                 0         -
-main.rs                                     3800                 0   100.00%         257                 0   100.00%        2508                 0   100.00%          66                 0   100.00%
+main.rs                                        3                 0   100.00%           1                 0   100.00%           3                 0   100.00%           0                 0         -
 migration.rs                                 690                 0   100.00%          34                 0   100.00%         306                 0   100.00%          30                 0   100.00%
 paths.rs                                     103                 0   100.00%          22                 0   100.00%          75                 0   100.00%           0                 0         -
 prompts.rs                                    70                 0   100.00%           7                 0   100.00%          39                 0   100.00%           0                 0         -
@@ -82,4 +83,4 @@ tui/render/modals/progress.rs                 99                 0   100.00%    
 tui/render/modals/settings_menu.rs            94                 0   100.00%           1                 0   100.00%          38                 0   100.00%           4                 0   100.00%
 update.rs                                    580                 0   100.00%          41                 0   100.00%         353                 0   100.00%          10                 0   100.00%
 ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-TOTAL                                      36948                 0   100.00%        2117                 0   100.00%       22255                 0   100.00%        1796                 0   100.00%
+TOTAL                                      34069                 0   100.00%        1932                 0   100.00%       20469                 0   100.00%        1854                 0   100.00%
