+++ ci/llvm-cov-receipts/windows/llvm-cov-report.sha256
@@ -1,1 +1,1 @@
-97d069713b6de3996f671ec0aa2bd8469d3b9ffed5a0d504eed8072a2e90001e  ci/llvm-cov-receipts/windows/llvm-cov-report.txt
+3ce28d92be56cb757bda35d919e603a4b5fc210d11a8d27d9aee93902fb88d23  ci/llvm-cov-receipts/windows/llvm-cov-report.txt
