+++ src/lib.rs
@@ -36,6 +36,7 @@ pub mod conversation;
 pub mod action;
 pub mod agent;
 pub mod app;
+pub mod cli;
 pub mod config;
 pub mod git;
 pub mod migration;
