+++ tests/unit.rs
@@ -1,5 +1,7 @@
 //! Unit tests migrated out of `src/`.
+#![cfg(feature = "test-support")]
 
 mod unit {
+    pub mod cli;
     pub mod support;
 }
